Move utils -> ext.utils
